+++ opensrp-chw/src/main/java/org/smartregister/chw/util/PNCVisitUtil.java
@@ -1,8 +1,12 @@
 package org.smartregister.chw.util;
 
 
+import android.content.Context;
+
+import org.apache.commons.lang3.StringUtils;
 import org.smartregister.chw.application.ChwApplication;
 import org.smartregister.chw.core.rule.PNCHealthFacilityVisitRule;
+import org.smartregister.chw.core.utils.Utils;
 
 import java.util.Date;
 
@@ -16,5 +20,11 @@ public class PNCVisitUtil {
         return visitRule;
     }
 
+    public static String getTranslatedValue(Context context, String name) {
+        if (StringUtils.isBlank(name))
+            return name;
+        String val = "pnc_" + name;
+        return Utils.getStringResourceByName(val, context);
+    }
 
 }
